Moved constants over to DuoUI singleton
